% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlim.R
\name{dlim}
\alias{dlim}
\title{Fit DLIM}
\usage{
dlim(
  y,
  x,
  modifiers,
  z = NULL,
  df_m,
  df_l,
  penalize = T,
  fit_fn = "gam",
  MaME = T
)
}
\arguments{
\item{y}{vector of response values}

\item{x}{matrix of exposure history}

\item{modifiers}{vector of modifying values}

\item{z}{matrix of z}

\item{df_m}{degrees of freedom for modifier basis}

\item{df_l}{degrees of freedom for exposure time basis}

\item{penalize}{True to penalize model}

\item{fit_fn}{specify "gam" to use the \code{gam} function for data sets that are not very large, and specify "bam" to use the \code{bam} function for data sets that are very large. Default will fit using \code{gam}.}

\item{MaME}{True to include modifier as linear main effect}
}
\value{
This function returns an object of class dlim
\item{cb}{cross-basis (matrix)}
\item{fit}{model object (gam)}
\item{modifiers}{modifying values (vector)}
\item{call}{model call}
}
\description{
Fit distributed lag interaction model
}
