% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_data.R
\docType{data}
\name{ex_data}
\alias{ex_data}
\title{Example data set}
\format{
List of response, exposure, modifiers, covariates
}
\source{
Simulated
}
\usage{
ex_data
}
\description{
Data set for examples
}
\examples{
data(ex_data) # lazy load
}
\keyword{datasets}
