% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\docType{data}
\name{exposure}
\alias{exposure}
\title{Exposure data set for simulation}
\format{
A data frame of 1000 rows and 37 columns
}
\source{
Data source??
}
\usage{
exposure
}
\description{
Data set of PM 2.5 exposure history for 1000 individuals over 37 weeks
}
\examples{
data(exposure) # lazy load
}
\keyword{datasets}
