% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{null_dlm_fit}
\alias{null_dlm_fit}
\title{Null DLM Fit}
\usage{
null_dlm_fit(fit, x, ...)
}
\arguments{
\item{fit}{dlim object}

\item{x}{exposure}

\item{...}{other args passed to [mgcv::gam]}
}
\value{
The refitted dlm fit
}
\description{
Return DLM fit
}
