% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative.R
\name{plot_cumulative}
\alias{plot_cumulative}
\title{Plot Distributed Lag Function}
\usage{
plot_cumulative(
  new_modifiers,
  mod_fit,
  dlm_fit = NULL,
  trans_fn = NULL,
  mod_name = NULL
)
}
\arguments{
\item{new_modifiers}{a vector of new modifier values for prediction (numeric)}

\item{mod_fit}{an object of class dlim (dlim)}

\item{dlm_fit}{a list containing a \code{crossbasis} object from the \pkg{dlnm} package as the first element and a DLM model object as the second element (list)}

\item{trans_fn}{if modifiers are transformed, specify back transformation function (character)}

\item{mod_name}{modifier name (character)}
}
\value{
This function returns ggplot
}
\description{
Plot estimated distributed lag function values from a DLIM object, can also compare those of a DLM
}
