% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sim_dlim.R
\name{predict.sim_dlim}
\alias{predict.sim_dlim}
\title{Simulated DLIM Predictions}
\usage{
\method{predict}{sim_dlim}(object, newdata = NULL, type = c("DLF", "CE", "response"))
}
\arguments{
\item{object}{an object of class dlim}

\item{newdata}{vector of modifiers for inference}

\item{type}{Type of prediction ("response" for predicted responses, "DLF" for the estimated distributed lag functions, "CE" for cumulative effects)}

\item{CI}{True to include confidence intervals (only for type="DLF" or type="CE")}
}
\value{

}
\description{
This function estimates cumulative and non-cumulative lag/modifier coefficients from a model in which the response is regressed on a cross-basis generated by the \code{cross_basis()} function.
}
\seealso{
\link[dlim]{predict.dlim}
}
