% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate Data}
\usage{
sim_data(
  x,
  L = NULL,
  modifiers,
  noise = 1,
  type = 2,
  SNR,
  ncovariates = 0,
  gamma = 1
)
}
\arguments{
\item{x}{a numeric time series vector of length n or matrix of lagged exposures for n individuals}

\item{L}{a numeric vector of length 1 containing the number of lag terms. This is required if \code{x} is vector, and is not used if \code{x} is a matrix.}

\item{modifiers}{vector of length n containing modifying values}

\item{noise}{a vector of length 1 containing the standard deviation for a normal distribution with mean 0 used to add noise to the simulated response values.}

\item{type}{a vector containing the number 1, 2, or 3 for simulation modification type: none, linear, complex. Default is 2}

\item{SNR}{value to magnify the signal of the true distributed lag functions, default is no magnification}

\item{ncovariates}{number of covariates to add to the model, numeric vector of length 1.}
}
\value{
This returns a list of 9 items:
\item{x}{a lagged exposure matrix. If \code{x} was a matrix, it is unchanged.}
\item{L}{a numeric vector o flength 1 containing the number of lag terms}
\item{modifiers}{the \code{modifiers} argument}
\item{y}{a numeric vector of length \code{nrow(x)} containing the perturbed simulated response values. obtained by adding a \code{N(0,noise)} random draw to \code{y}}
\item{betas}{a matrix containing true coefficients for each lag/modifier combination, with each row representing a lag and each column a modifier}
\item{betas_cumul}{a numeric vector of length \code{L+1} containing cumululative true coefficients for the lag terms, summed over modifiers}
\item{sim_type}{simulation type 1, 2, or 3 (numeric)}
\item{Z}{covariates (matrix)}
\item{gammas}{true coefficients for the covariates (numeric)}
}
\description{
Simulate data to use with the \pkg{dlim} package. There are different scenarios to choose for simulation.
}
