% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dlf.R
\name{sim_dlf}
\alias{sim_dlf}
\title{Simulate Distributed Lag Functions}
\usage{
sim_dlf(L, modifiers, type)
}
\arguments{
\item{L}{number of lags minus 1}

\item{type}{simulation type: 1 is no modification, 2 is linear scale modification, 3 is logistic shift modification, 4 is types 2 and 3 combined}

\item{M}{vector of modifiers}
}
\description{
generate true distributed lag function values for a given type of simulation
}
\seealso{
\link[dlim]{sim_data}
}
