% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{standard_dlm}
\alias{standard_dlm}
\title{Standard DLM Fit}
\usage{
standard_dlm(
  x,
  arg_var = list(fun = "lin"),
  arg_lag = list(fun = "ps", df = 20),
  z,
  modifiers,
  y,
  family,
  method,
  ...
)
}
\arguments{
\item{x}{exposure}

\item{arg_var}{see [dlnm::crossbasis]}

\item{arg_lag}{see [dlnm::crossbasis]}

\item{z}{covariate matrix}

\item{modifiers}{modifier}

\item{y}{outcome}

\item{family}{family object or family}

\item{method}{passed to [mgcv::gam], REML or ML or GCV etc}

\item{...}{other args passed to [mgcv::gam]}

\item{fit}{dlim object}
}
\value{
The refitted dlm fit
}
\description{
Return DLM fit
}
