% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dlim.R
\name{summary.dlim}
\alias{summary.dlim}
\title{Summarizing DLIM}
\usage{
\method{summary}{dlim}(x)
}
\arguments{
\item{x}{a \code{dlim} object}
}
\value{
This function returns a summary for an object of class \code{dlim}
}
\description{
prints summary of object of class \code{dlim}
}
