% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{model_comparison}
\alias{model_comparison}
\title{Model Comparison}
\usage{
model_comparison(fit, null = "DLM", x, B, conf.level = 0.95)
}
\arguments{
\item{fit}{dlim object (must be fit with REML)}

\item{null}{"DLM", "linear" to specify the null model}

\item{x}{exposure}

\item{B}{number of bootstrap samples}

\item{conf.level}{The confidence level (class "\code{numeric}")}
}
\value{
The function returns a decision to either reject or fail to reject the null model
}
\description{
Compare models to test for interaction
}
\seealso{
Type \code{vignette('dlimOverview')} for a detailed description.

\link[dlim]{dlim}
}
