% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DLF.R
\name{plot_DLF}
\alias{plot_DLF}
\title{Plot Cumulative Effects}
\usage{
plot_DLF(
  new_modifiers,
  mod_fit,
  mod_name,
  dlm_fit = NULL,
  plot_by,
  time_pts = NULL,
  trans_fn = NULL
)
}
\arguments{
\item{new_modifiers}{a vector of new modifier values for prediction (numeric)}

\item{mod_fit}{an object of class dlim (dlim)}

\item{mod_name}{modifier name (character)}

\item{dlm_fit}{a list containing a \code{crossbasis} object from the \pkg{dlnm} package as the first element and a DLM model object as the second element (list)}

\item{plot_by}{choose to create plots for particular modifier values, "modifier", or particular time points, "time", (character)}

\item{time_pts}{a set of time points if plotting by time (numeric)}

\item{trans_fn}{if modifiers are transformed, specify back transformation function (character)}
}
\value{
This function returns ggplot
}
\description{
Plot estimated cumulative effects from a DLIM object, can also compare estimated cumulative effects between a DLM and DLIM
}
