% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dlim.R
\name{print.dlim}
\alias{print.dlim}
\title{Print DLIM Information}
\usage{
\method{print}{dlim}(x, ...)
}
\arguments{
\item{x}{a \code{dlim} object}

\item{...}{further arguments passed to or from other methods}
}
\value{
This function returns information about an object of class \code{dlim}
}
\description{
prints information about an object of class \code{dlim}
}
\seealso{
Type \code{vignette('dlimOverview')} for a detailed description.
}
