% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dlf.R
\name{sim_dlf}
\alias{sim_dlf}
\title{Simulate Distributed Lag Functions}
\usage{
sim_dlf(L, modifiers, type, mag = 1)
}
\arguments{
\item{L}{number of lags minus 1}

\item{type}{simulation type: 1 is no modification, 2 is linear modification, 3 is complex modification, 4 is stronger complex modification}

\item{mag}{value to magnify the signal of the true distributed lag function, default is no magnification}

\item{M}{vector of modifiers}
}
\description{
generate true distributed lag function values for a given type of simulation
}
\seealso{
\link[dlim]{sim_data}
}
