% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dlf.R
\name{sim_dlf}
\alias{sim_dlf}
\title{Simulate Distributed Lag Functions}
\usage{
sim_dlf(L, modifiers, type)
}
\arguments{
\item{L}{Number of lags minus 1}

\item{modifiers}{Vector of modifiers}

\item{type}{Effect modification simulation type: 1 is no modification, 2 is linear scale modification, 3 is non-linear shift modification, 4 is types 2 and 3 combined}
}
\value{
This function returns the true distributed lag function values (class "\code{numeric}")
}
\description{
generate true distributed lag function values for a given type of simulation
}
\seealso{
\link[dlim]{sim_data}

Type \code{vignette('dlimOverview')} for a detailed description.
}
