% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dlim.R
\name{sim_dlim}
\alias{sim_dlim}
\title{Fit DLIM for simulation}
\usage{
sim_dlim(
  data,
  df_m,
  df_l,
  penalize = T,
  fit_dlm = F,
  model_type = "standard",
  ...
)
}
\arguments{
\item{data}{output from sim_data()}

\item{df_m}{degrees of freedom for modifiers}

\item{df_l}{degrees of freedom for lags}

\item{penalize}{True to penalize model}

\item{fit_dlm}{True to additionally fit dlm for comparison}

\item{model_type}{"linear" for a DLIM with linear interaction, "quadratic" for a DLIM with quadratic interaction, "standard" for a DLIM with splines}

\item{...}{arguments to pass to model fitting function}
}
\value{
This function returns an object of class dlim
}
\description{
Fit DLIM for simulation
}
\seealso{
\link[dlim]{dlim}
}
